# api/collaboration/openapi.yaml
openapi: 3.1.0
info:
  title: AphelionDMM Collaboration Control API
  version: 1.0.0
  description: Version 1 lifecycle and snapshot contract. Filesystem paths and process commands are never accepted.
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://127.0.0.1
  - url: https://collaboration.example
paths:
...
  /v1/hosted/sessions:
    get:
      summary: List active Community sessions or enabled owned/admitted sessions
      operationId: listHostedSessions
      security: [{ bearerToken: [] }]
      parameters:
        - { name: scope, in: query, required: true, schema: { type: string, enum: [community, mine] } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 50 } }
        - { name: cursor, in: query, schema: { type: string, maxLength: 128 } }
        - { name: include, in: query, description: 'Set to repository only after the capabilities response advertised X-Aphelion-Repository-Descriptor. Without it rows never contain repository, so strict older clients keep decoding.', schema: { type: string, enum: [repository] } }
      responses:
        '200':
          description: Stable ID cursor; activity is a point-in-time observation; compatibility remains unchecked
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [sessions]
                properties:
                  sessions: { type: array, maxItems: 100, items: { $ref: '#/components/schemas/HostedSessionSummary' } }
                  next_cursor: { type: string, maxLength: 128 }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
        '501': { $ref: '#/components/responses/Error' }
        '503': { $ref: '#/components/responses/Error' }
    post:
      summary: Create a durable hosted collaboration session
      operationId: createHostedSession
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHostedSessionRequest' }
      responses:
        '201':
          description: Durable hosted session created with the authenticated identity as owner
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedSessionResponse' }
        '400': { $ref: '#/components/responses/Error' }
        '401': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
        '413': { $ref: '#/components/responses/Error' }
...
  /v1/sessions/{session_id}/hosted-invitations:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Create a one-use hosted invitation
      operationId: createHostedInvitation
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateHostedInvitationRequest' }
      responses:
        '201':
          description: One-use hosted invitation. The raw token is returned only in this response.
          headers:
            Cache-Control: { schema: { type: string, const: no-store } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedInvitation' }
        '403': { $ref: '#/components/responses/Error' }
  /v1/sessions/{session_id}/hosted-invitations/redeem:
    parameters: [{ $ref: '#/components/parameters/SessionID' }]
    post:
      summary: Redeem a one-use hosted invitation
      operationId: redeemHostedInvitation
      security: [{ bearerToken: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/RedeemHostedInvitationRequest' }
      responses:
        '200':
          description: Hosted membership created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/HostedMembership' }
        '401': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    launchToken: { type: http, scheme: bearer, description: Single-use embedded launch token }
    bearerToken: { type: http, scheme: bearer }
  parameters:
    SessionID:
      name: session_id
      in: path
      required: true
      schema: { type: string, minLength: 1, maxLength: 128 }
  responses:
    Error:
      description: Bounded stable error
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    SHA256:
      type: string
      pattern: '^[0-9a-f]{64}$'
...
    HostedSessionResponse:
      type: object
      additionalProperties: false
      required: [session_id, document_id, revision, map_hash]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        document_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        map_hash: { $ref: '#/components/schemas/SHA256' }
    CreateHostedSessionRequest:
      type: object
      additionalProperties: false
      required: [snapshot]
      properties:
        snapshot: { $ref: '#/components/schemas/Snapshot' }
        bulk_edits: { type: boolean, default: false }
        visibility: { type: string, enum: [private, community], default: private }
        title: { type: string, maxLength: 128, description: Maximum 128 UTF-8 bytes; labels cannot contain controls or absolute paths }
        map_label: { type: string, maxLength: 128 }
        environment_label: { type: string, maxLength: 128 }
        repository:
          $ref: '#/components/schemas/RepositoryDescriptor'
    RepositoryDescriptor:
      type: object
      additionalProperties: false
      description: Optional, additive host repository identity. Carries no paths, remote URLs or credentials. Held in service memory only, so a restarted service lists recovered sessions without it; absence means unknown. Receivers validate again and treat failures as unknown.
      required: [dme_name, environment_hash]
      properties:
        dme_name: { type: string, minLength: 5, maxLength: 128, pattern: '^[^/\\:]+\.[dD][mM][eE]$', description: DME file name only }
        environment_hash: { $ref: '#/components/schemas/SHA256' }
        git_branch: { type: string, minLength: 1, maxLength: 200, pattern: '^[A-Za-z0-9._/][A-Za-z0-9._/-]*$', description: Omitted for a detached HEAD or non-git checkout; requires git_commit }
        git_commit: { type: string, pattern: '^[0-9a-f]{40}$', description: Omitted for a non-git checkout }
...
    HostedSessionSummary:
      type: object
      additionalProperties: false
      required: [session_id, visibility, title, map_label, environment_label, owner_display_name, participants, available]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        visibility: { type: string, enum: [private, community] }
        title: { type: string, maxLength: 128 }
        map_label: { type: string, maxLength: 128 }
        environment_label: { type: string, maxLength: 128 }
        owner_display_name: { type: string, maxLength: 128 }
        participants: { type: integer, minimum: 0 }
        available: { type: boolean }
        repository:
          $ref: '#/components/schemas/RepositoryDescriptor'
...
    CreateHostedInvitationRequest:
      type: object
      additionalProperties: false
      required: [role]
      properties:
        role: { type: string, enum: [viewer, editor] }
    HostedInvitation:
      type: object
      additionalProperties: false
      required: [token, role, expires_at]
      properties:
        token: { type: string, minLength: 1, maxLength: 128 }
        role: { type: string, enum: [viewer, editor] }
        expires_at: { type: string, format: date-time }
    RedeemHostedInvitationRequest:
      type: object
      additionalProperties: false
      required: [token]
      properties:
        token: { type: string, minLength: 1, maxLength: 128 }
    HostedMembership:
      type: object
      additionalProperties: false
      required: [session_id, actor_id, role]
      properties:
        session_id: { type: string, minLength: 1, maxLength: 128 }
        actor_id: { type: string, format: uuid }
        role: { type: string, enum: [viewer, editor, owner] }
...
    Snapshot:
      type: object
      additionalProperties: false
      required: [protocol_version, schema_version, document_id, revision, environment_hash, max_x, max_y, max_z, tiles]
      properties:
        protocol_version: { type: integer, const: 1 }
        schema_version: { type: integer, const: 1 }
        document_id: { type: string, format: uuid }
        revision: { type: integer, minimum: 0 }
        environment_hash: { $ref: '#/components/schemas/SHA256' }
        max_x: { type: integer, minimum: 1, maximum: 4096 }
        max_y: { type: integer, minimum: 1, maximum: 4096 }
        max_z: { type: integer, minimum: 1, maximum: 4096 }
        tiles: { type: array, maxItems: 16777216, items: { type: object } }
...
    Error:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string, minLength: 1, maxLength: 128 }
        message: { type: string, minLength: 1, maxLength: 512 }
        current_revision: { type: integer, minimum: 0 }
        current_map_hash: { $ref: '#/components/schemas/SHA256' }
...